[FTheoryTools] * Attributes, properties, setters of literature models moved to AbstractFTheoryModels code * Extended by yet more setters and adders
