+++ experimental/FTheoryTools/src/exports.jl
@@ -8,7 +8,19 @@ export WeierstrassModel
 export _blowup_global
 export _blowup_global_sequence
 export _is_nontrivial
+export add_associated_literature_model
+export add_generating_section
+export add_journal_report_number
+export add_model_parameter
+export add_paper_author
+export add_paper_buzzword
+export add_related_literature_model
 export add_resolution
+export add_resolution_generating_section
+export add_resolution_zero_section
+export add_weighted_resolution
+export add_weighted_resolution_generating_section
+export add_weighted_resolution_zero_section
 export ambient_space
 export analyze_fibers
 export arxiv_doi
@@ -96,9 +108,42 @@ export resolution_zero_sections
 export resolutions
 export resolve
 export sample_toric_variety
-export set_description
+export set_arxiv_doi
+export set_arxiv_id
+export set_arxiv_link
+export set_arxiv_model_equation_number
+export set_arxiv_model_page
+export set_arxiv_model_section
+export set_arxiv_version
+export set_associated_literature_models
+export set_generating_sections
 export set_global_tate_model
+export set_journal_doi
+export set_journal_link
+export set_journal_model_equation_number
+export set_journal_model_page
+export set_journal_model_section
+export set_journal_pages
+export set_journal_report_numbers
+export set_journal_volume
+export set_journal_year
+export set_literature_identifier
+export set_model_description
+export set_model_description
+export set_model_parameters
+export set_paper_authors
+export set_paper_buzzwords
+export set_paper_description
+export set_paper_title
+export set_related_literature_models
+export set_resolution_generating_sections
+export set_resolution_zero_sections
+export set_resolutions
 export set_weierstrass_model
+export set_weighted_resolution_generating_sections
+export set_weighted_resolution_zero_sections
+export set_weighted_resolutions
+export set_zero_section
 export singular_loci
 export su5_tate_model_over_arbitrary_3d_base
 export su5_weierstrass_model_over_arbitrary_3d_base
